# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for tracking daily activities, completions, and rewards.",
        "title": "Routine Series API",
        "contact": {},
        "version": "0.1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/completions": {
            "get": {
                "description": "Returns completions for an activity within a date range.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "completions"
                ],
                "summary": "List completions",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Activity ID",
                        "name": "activity_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Start date (YYYY-MM-DD)",
                        "name": "from",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "End date (YYYY-MM-DD)",
                        "name": "to",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/completion.Completion"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/completions/toggle": {
            "post": {
                "description": "Creates a completion mark for the given date if none exists; deletes it if one already exists.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "completions"
                ],
                "summary": "Toggle completion",
                "parameters": [
                    {
                        "description": "Completion data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/completion.ToggleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/completion.ToggleResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/reward-issues": {
            "get": {
                "description": "Returns paginated reward issues for an activity, newest first.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rewards"
                ],
                "summary": "List reward issues",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Activity ID",
                        "name": "activity_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Limit (default 20)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset (default 0)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/reward.PaginatedResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Records a reward issuance for an activity.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rewards"
                ],
                "summary": "Create reward issue",
                "parameters": [
                    {
                        "description": "Reward data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/reward.CreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/reward.RewardIssue"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/reward-issues/{id}": {
            "delete": {
                "description": "Deletes a reward issue by ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rewards"
                ],
                "summary": "Delete reward issue",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Reward Issue ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Updates fields of a reward issue by ID. All fields are optional; at least one must be provided.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rewards"
                ],
                "summary": "Update reward issue",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Reward Issue ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fields to update",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/reward.UpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/reward.RewardIssue"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "api.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "completion.Completion": {
            "type": "object",
            "properties": {
                "activity_id": {
                    "type": "integer"
                },
                "date": {
                    "description": "YYYY-MM-DD",
                    "type": "string",
                    "example": "2026-08-12"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
        "completion.ToggleRequest": {
            "type": "object",
            "properties": {
                "activity_id": {
                    "type": "integer",
                    "example": 1
                },
                "date": {
                    "description": "YYYY-MM-DD",
                    "type": "string",
                    "example": "2026-08-12"
                }
            }
        },
        "completion.ToggleResponse": {
            "type": "object",
            "properties": {
                "completion": {
                    "$ref": "#/definitions/completion.Completion"
                },
                "created": {
                    "type": "boolean"
                }
            }
        },
...
        "reward.CreateRequest": {
            "type": "object",
            "properties": {
                "activity_id": {
                    "type": "integer",
                    "example": 1
                },
                "amount": {
                    "type": "number",
                    "example": 100
                },
                "currency": {
                    "type": "string",
                    "example": "₽"
                },
                "date": {
                    "description": "YYYY-MM-DD",
                    "type": "string",
                    "example": "2026-08-12"
                }
            }
        },
        "reward.PaginatedResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/reward.RewardIssue"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "reward.RewardIssue": {
            "type": "object",
            "properties": {
                "activity_id": {
                    "type": "integer"
                },
                "amount": {
                    "type": "number"
                },
                "currency": {
                    "type": "string"
                },
                "date": {
                    "description": "YYYY-MM-DD",
                    "type": "string",
                    "example": "2026-08-12"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
        "reward.UpdateRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number",
                    "example": 150
                },
                "currency": {
                    "type": "string",
                    "example": "₽"
                },
                "date": {
                    "description": "YYYY-MM-DD",
                    "type": "string",
                    "example": "2026-08-12"
                }
            }
        },
...
    }
}